requirements.txt
asgiref==3.7.2
baserowapi==0.1.0b1
certifi==2024.2.2
charset-normalizer==3.3.2
databind.core==4.4.2
databind.json==4.4.2
Deprecated==1.2.14
Django==5.0.2
django-dotenv==1.4.2
djangorestframework==3.14.0
idna==3.6
nr-date==2.1.0
nr-stream==1.1.5
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
typeapi==2.1.2
typing_extensions==4.6.3
urllib3==2.2.1
wrapt==1.16.0
